firstgame 정리
정리

+++ MyFirstGame/MyFirstGame/Environment.h
@@ -10,8 +10,8 @@ class CEnvironment
 
 	//SDL_Event* events;
 	CSDL_setup* csdl_setup;
-	int* CameraX;
-	int* CameraY;
+	float* CameraX;
+	float* CameraY;
 
 	//ÀÌ¹ÌÁö
 	CSprite* background[4][7];
